pyproject.toml
[project]
name = "stroke-unit"
version = "0.1.0"
description = "A Django project for stroke unit management"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-widget-tweaks>=1.5.0,<2.0.0",
    "requests>=2.31.0",
    "faker>=19.13.0",
    # Production dependencies
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "dj-database-url>=2.1.0",
    "python-decouple>=3.8"
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.2.0",
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "coverage>=7.3.2"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
